ver. 0.21.1

+++ main.cpp
@@ -8,7 +8,7 @@ int main(int argc, char *argv[]) {
     int seed;
     if (argc == 2) seed = atoi(argv[1]);
     else seed = time(NULL);
-    srand(seed);
+    srand((unsigned int) seed);
 
     Battle navalBattle;
